Invoice Parsing System/requirements.txt
requests
pyyaml
public_ip
boto3
